requirements.dev.txt
# Core Django and Web
Django>=5.2.7
djangorestframework>=3.16.1
django-cors-headers>=4.9.0
django-environ>=0.12.0
python-dotenv>=1.1.1
gunicorn>=23.0.0
whitenoise>=6.11.0
dj-static>=0.0.6

# Database
psycopg2-binary>=2.9.11
dj-database-url>=3.0.1
sqlalchemy>=2.0.44
alembic>=1.17.0

# Async & Task Queue
celery>=5.5.3
celery[redis]>=5.5.3
flower>=2.0.1
django-celery-beat>=2.8.1
django-celery-results>=2.6.0
redis>=5.2.0,<5.3.0  # Compatible with celery 5.5.3 (requires <=5.2.1)
django-redis>=6.0.0

# API & Auth
fastapi>=0.115.0  # Legacy API routes (being migrated to Django)
uvicorn>=0.32.0  # ASGI server for FastAPI
passlib>=1.7.4  # Password hashing for FastAPI auth
python-jose>=3.5.0  # JWT token handling for FastAPI
djangorestframework-simplejwt>=5.5.1
drf-spectacular>=0.28.0
django-filter>=25.2
user-agents>=2.2.0

# Caching and real-time
websocket-client>=1.9.0
websockets>=15.0.1

# API and async
aiohttp>=3.13.1

# Data Science & ML
pandas>=2.2.0  # 2.3.x requires Python 3.11+
numpy>=1.26.0,<2.0  # 2.x has compatibility issues
scipy>=1.13.0
scikit-learn>=1.5.0
hmmlearn>=0.3.2
joblib>=1.4.0
xgboost>=2.1.0
lightgbm>=4.5.0

# Technical Analysis
TA-Lib>=0.6.8

# Crypto/Finance
ccxt>=4.4.0
yfinance>=0.2.48

# AI/ML Libraries
autots>=0.6.15
prophet>=1.1.5
statsmodels>=0.14.2

# LLM & Chatbot
openai>=1.54.0
google-generativeai>=0.8.3
anthropic>=0.39.0

# RAG & Document Processing
langchain>=0.3.0
langchain-community>=0.3.0
langchain-openai>=0.2.0
faiss-cpu>=1.8.0
chromadb>=0.5.0
pgvector>=0.3.5
tiktoken>=0.8.0

# Local LLM Support (CUDA)
ollama>=0.4.0
transformers>=4.46.0
accelerate>=1.0.0
bitsandbytes>=0.44.0

# Document Parsing
pypdf>=5.1.0
PyMuPDF>=1.24.0
python-docx>=1.1.2
openpyxl>=3.1.5

# Optimization
optuna>=4.0.0
optuna-dashboard>=0.16.0

# Visualization
matplotlib>=3.9.0
seaborn>=0.13.2
plotly>=5.24.0
streamlit>=1.40.0

# Utils
requests>=2.32.3
pytz>=2024.2
python-dateutil>=2.9.0.post0
psutil>=6.1.0

# Monitoring & Observability
prometheus-client>=0.21.0
django-prometheus>=2.3.1

# Testing
pytest>=8.3.0
pytest-django>=4.9.0
pytest-cov>=6.0.0
pytest-asyncio>=0.24.0
pytest-mock>=3.14.0
faker>=30.8.0
factory-boy>=3.3.1

# Code Quality
black>=24.10.0
flake8>=7.1.0
isort>=5.13.0
pylint>=3.3.0
mypy>=1.13.0

# Monitoring & Logging
loguru>=0.7.2
sentry-sdk>=2.18.0
django-debug-toolbar>=4.4.0
django-extensions>=3.2.3

# Security
django-ratelimit>=4.1.0
django-axes>=8.0.0
pip-audit>=2.7.0

# Development
ipython>=8.29.0
jupyter>=1.1.1
notebook>=7.2.0
